+++ changelog.md
@@ -1,2 +1,1 @@
-- Update JEI and WAILA
-- Fixed canister crash
+- Fixed fuel station crash